requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-filter==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
factory-boy==2.12.0
Faker==4.1.1
idna==2.9
importlib-metadata==1.6.1
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
more-itertools==8.4.0
openapi-codec==1.3.2
packaging==20.4
Pillow==5.0.0
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
